feat(core): compute all workspace files incrementally
