--- packages/workspace/src/core/file-graph/index.ts
@@ -1,1 +1,0 @@
-export * from './project-file-map';
