--- packages/workspace/src/core/file-graph/project-file-map.ts
@@ -1,36 +1,0 @@
-import { FileData, ProjectFileMap } from '@nrwl/devkit';
-
-export function createProjectFileMap(
-  workspaceJson: any,
-  files: FileData[]
-): ProjectFileMap {
-  const fileMap: ProjectFileMap = {};
-  const seen = new Set();
-  // Sorting here so `apps/client-e2e` comes before `apps/client` and has
-  // a chance to match prefix first.
-  const sortedProjects = Object.keys(workspaceJson.projects).sort((a, b) => {
-    const projectA = workspaceJson.projects[a];
-    const projectB = workspaceJson.projects[b];
-    if (!projectA.root) return -1;
-    if (!projectB.root) return -1;
-    return projectA.root.length > projectB.root.length ? -1 : 1;
-  });
-
-  sortedProjects.forEach((projectName) => {
-    const p = workspaceJson.projects[projectName];
-    fileMap[projectName] = [];
-    files.forEach((f) => {
-      if (seen.has(f.file)) {
-        return;
-      }
-      if (f.file.startsWith(p.root || p.sourceRoot)) {
-        fileMap[projectName].push(f);
-        seen.add(f.file);
-      }
-    });
-  });
-
-  return fileMap;
-}
-
-export { ProjectFileMap } from '@nrwl/devkit';
