+++ packages/workspace/src/core/project-graph/build-dependencies/explicit-package-json-dependencies.ts
@@ -4,8 +4,6 @@ import {
   joinPathFragments,
   parseJson,
   ProjectFileMap,
-  ProjectGraphBuilder,
-  ProjectGraphProcessorContext,
   Workspace,
 } from '@nrwl/devkit';
