+++ packages/workspace/src/core/project-graph/index.ts
@@ -1,8 +1,6 @@
 export {
-  createProjectGraph,
   createProjectGraphAsync,
   readCachedProjectGraph,
-  createProjectGraphInCurrentProcess,
 } from './project-graph';
 export * from './project-graph-models';
 export * from './operators';
